8088 and CGA now have more accurate timings (but 8088 is still not cycle exact); Added RTL 8029AS (NE2000 PCI); Applies RichardG's SLiRP TFTP patch; Added missing 8088 instruction aliases and SALC instruction; Possibly fixed some PIIX PCI bugs; Added NVidia Riva TNT emulation (not yet used); PIT is now more accurate; ATAPI TEST UNIT READY command now sets discchanged to 1; SYSENTER, SYSEXIT, SYSCALL, SYSRET, and LOADALL386 now end the current recompiler block; Network type (PCap, SLiRP) is now configurable inside the emulator and per adapter; Added support for optional Etherboot network ROM with the RTL 8029AS; Possibly other changes I am not aware of; Added a makefile for network logging.

+++ PCem/sound_opl.c
@@ -12,7 +12,8 @@ uint8_t opl2_read(uint16_t a, void *priv)
 {
         opl_t *opl = (opl_t *)priv;
 
-        cycles -= (int)(isa_timing * 8);
+        // cycles -= (int)(isa_timing * 8);
+        cycles -= (isa_timing * 8);
         return opl_read(0, a);
 }
 void opl2_write(uint16_t a, uint8_t v, void *priv)
@@ -27,7 +28,8 @@ uint8_t opl2_l_read(uint16_t a, void *priv)
 {
         opl_t *opl = (opl_t *)priv;
 
-        cycles -= (int)(isa_timing * 8);
+        // cycles -= (int)(isa_timing * 8);
+        cycles -= (isa_timing * 8);
         return opl_read(0, a);
 }
 void opl2_l_write(uint16_t a, uint8_t v, void *priv)
@@ -41,7 +43,8 @@ uint8_t opl2_r_read(uint16_t a, void *priv)
 {
         opl_t *opl = (opl_t *)priv;
 
-        cycles -= (int)(isa_timing * 8);
+        // cycles -= (int)(isa_timing * 8);
+        cycles -= (isa_timing * 8);
         return opl_read(1, a);
 }
 void opl2_r_write(uint16_t a, uint8_t v, void *priv)
@@ -55,7 +58,8 @@ uint8_t opl3_read(uint16_t a, void *priv)
 {
         opl_t *opl = (opl_t *)priv;
 
-        cycles -= (int)(isa_timing * 8);
+        // cycles -= (int)(isa_timing * 8);
+        cycles -= (isa_timing * 8);
         return opl_read(0, a);
 }
 void opl3_write(uint16_t a, uint8_t v, void *priv)
